Refactor archivematicaselenium.py to amuser
Refactors archivematicaselenium.py to the package amuser/.
What was archivematicaselenium.py is now broken up into four high-level
abilities:

1. Browser ability (wraps Selenium)
2. API ability (wraps Python's Requests library)
3. Command Line ability (wraps Python's pexpect and subprocess libraries for
   making SSH and scp calls)
4. METS ability (wraps lxml)
