+++ .gitignore
@@ -12,5 +12,6 @@ test-files/
 __pycache__
 behave-tutorial/
 reports/
+data/
 archivematicaselenium.log
 steps.log
