+++ amuser/am_browser_preservation_planning_ability.py
@@ -1,0 +1,201 @@
+"""Archivematica Browser Preservation Planning Ability"""
+
+from selenium.webdriver.common.keys import Keys
+from selenium.common.exceptions import NoSuchElementException
+
+from . import selenium_ability
+from . import utils
+
+
+LOGGER = utils.LOGGER
+
+
+class ArchivematicaBrowserPreservationPlanningAbility(
+        selenium_ability.ArchivematicaSeleniumAbility):
+    """Archivematica Browser Ability: the ability of an Archivematica user to
+    use a browser to interact with Archivematica. A class for using Selenium to
+    interact with a live Archivematica instance.
+    """
+
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+
+    def navigate_to_preservation_planning(self):
+        self.navigate(self.get_preservation_planning_url())
+
+    def navigate_to_normalization_rules(self):
+        self.navigate(self.get_normalization_rules_url())
+
+    def search_rules(self, search_term):
+        search_input_el = self.driver.find_element_by_css_selector(
+            '#DataTables_Table_0_filter input')
+        search_input_el.send_keys(search_term)
+
+    def click_first_rule_replace_link(self):
+        """Click the "replace" link of the first rule in the FPR rules table
+        visible on the page.
+        """
+        for a_el in self.driver.find_elements_by_tag_name('a'):
+            if a_el.text.strip() == 'Replace':
+                a_el.click()
+                break
+
+    def wait_for_rule_edit_interface(self):
+        self.wait_for_presence('#id_f-purpose')
+
+    def set_fpr_command(self, command_name):
+        command_select_el = self.driver.find_element_by_id('id_f-command')
+        command_select_el.click()
+        command_select_el.send_keys(command_name)
+        command_select_el.send_keys(Keys.RETURN)
+
+    def save_fpr_command(self):
+        command_select_el = self.driver.find_element_by_css_selector(
+            'input[type=submit]')
+        command_select_el.click()
+        self.wait_for_presence('#DataTables_Table_0')
+
+    def change_normalization_rule_command(self, search_term, command_name):
+        """Edit the FPR normalization rule that uniquely matches
+        ``search_term`` so that its command is the one matching
+        ``command_name``.
+        """
+        self.navigate_to_normalization_rules()
+        self.search_rules(search_term)
+        self.click_first_rule_replace_link()
+        self.wait_for_rule_edit_interface()
+        self.set_fpr_command(command_name)
+        self.save_fpr_command()
+
+    def navigate_to_first_policy_check_validation_command(self):
+        """Find the first policy check validation command and navigate to it.
+        Assumes that we are at the validation commands URL and that there is at
+        least one policy check validation command in this AM. Returns a list of
+        existing policy check command descriptions.
+        """
+        policy_command_url = None
+        policy_command_descriptions = []
+        commands_table_el = self.driver.find_element_by_id(
+            'DataTables_Table_0')
+        for row_el in commands_table_el.find_elements_by_tag_name('tr'):
+            try:
+                anchor_el = row_el.find_element_by_tag_name('a')
+            except NoSuchElementException:
+                pass
+            else:
+                if anchor_el.text.strip().startswith('Check against policy '):
+                    policy_command_url = anchor_el.get_attribute('href')
+                    policy_command_descriptions.append(anchor_el.text.strip())
+        if policy_command_url:
+            self.navigate(policy_command_url)
+            return policy_command_descriptions
+        else:
+            return []
+
+    def ensure_fpr_policy_check_command(self, policy_file, policy_path):
+        """Ensure there is an FPR validation command that checks a file against
+        the MediaConch policy ``policy_file``.
+        """
+        self.navigate(self.get_validation_commands_url())
+        existing_policy_command_descriptions = \
+            self.navigate_to_first_policy_check_validation_command()
+        description = get_policy_command_description(policy_file)
+        if description in existing_policy_command_descriptions:
+            # This policy command already exists; no need to re-create it.
+            return
+        policy_command = self.get_policy_command(policy_file, policy_path)
+        self.save_policy_check_command(policy_command, description)
+
+    def get_policy_command(self, policy_file, policy_path):
+        """Return a string representing a policy check validation command that
+        references the policy file ``policy_file``. Assumes that we are
+        viewing an existing validation-via-mediaconch-policy command.
+        """
+        # Get the text of the command.
+        policy_command = None
+        next_el = False
+        for el in self.driver.find_element_by_tag_name('dl')\
+                             .find_elements_by_css_selector('*'):
+            if next_el:
+                policy_command = el.find_element_by_tag_name('pre')\
+                                        .text.strip()
+                break
+            if el.text.strip() == 'Command':
+                next_el = True
+        # Insert our policy file name into the command text.
+        lines = []
+        with open(policy_path) as filei:
+            policy_lines = filei.read().splitlines()
+        for line in policy_command.splitlines():
+            if line.strip().startswith('POLICY = """'):
+                lines.append(line)
+                lines += policy_lines
+            elif line.strip() == 'POLICY_NAME = \'\'':
+                lines.append('POLICY_NAME = \'{}\''.format(policy_file))
+            else:
+                lines.append(line)
+        return '\n'.join(lines)
+
+    def save_policy_check_command(self, policy_command, description):
+        """Create and save a new FPR command using the string
+        ``policy_command``."""
+        self.navigate(self.get_create_command_url())
+        self.driver.find_element_by_id('id_tool').send_keys('MediaConch')
+        self.driver.find_element_by_id('id_description').send_keys(description)
+        js_script = ('document.getElementById("id_command").value ='
+                     ' `{}`;'.format(policy_command))
+        self.driver.execute_script(js_script)
+        self.driver.find_element_by_id('id_script_type').send_keys('Python')
+        self.driver.find_element_by_id('id_command_usage').send_keys(
+            'Validation')
+        self.driver.find_element_by_css_selector('input[type=submit]').click()
+
+    def ensure_fpr_rule(self, purpose, format_, command_description):
+        """Ensure that there is a new FPR rule with the purpose, format and
+        command description given in the params.
+        Note that the ``format_`` param is assumed to be in the format that the
+        /fpr/fprule/create/ expects, i.e., a colon-delimited triple like
+        'Audio: Broadcast WAVE: Broadcast WAVE 1'.
+        """
+        if self.fpr_rule_already_exists(purpose, format_, command_description):
+            return
+        self.navigate(self.get_create_rule_url())
+        self.driver.find_element_by_id('id_f-purpose').send_keys(purpose)
+        self.driver.find_element_by_id('id_f-format').send_keys(format_)
+        self.driver.find_element_by_id('id_f-command').send_keys(
+            command_description)
+        self.driver.find_element_by_css_selector('input[type=submit]').click()
+
+    def fpr_rule_already_exists(self, purpose, format_, command_description):
+        """Return ``True`` if an FPR rule already exists with the purpose,
+        format and command description given in the params; ``False`` otherwise.
+        """
+        self.navigate(self.get_rules_url())
+        self.search_for_fpr_rule(purpose, format_, command_description)
+        info_el = self.driver.find_element_by_id('DataTables_Table_0_info')
+        if info_el.text.strip().startswith('Showing 0 to 0 of 0 entries'):
+            return False
+        return True
+
+    def search_for_fpr_rule(self, purpose, format_, command_description):
+        """Search for an FPR rule with the supplied purpose, format and command
+        description. Uses the FPR asynchronous search input.
+        """
+        terse_format = format_.split(':')[2].strip()
+        search_term = '{} {} {}'.format(purpose, terse_format,
+                                        command_description)
+        self.search_rules(search_term)
+
+    def ensure_fpr_rule_enabled(self, purpose, format_, command_description):
+        self.navigate(self.get_rules_url())
+        self.search_for_fpr_rule(purpose, format_, command_description)
+        info_el = self.driver.find_element_by_id('DataTables_Table_0_info')
+        if info_el.text.strip().startswith('Showing 0 to 0 of 0 entries'):
+            return
+        # TODO: click the "Enable" link. But we have to make sure there is only
+        # one matching rule that needs enabling. Not sure at this point whether
+        # this action is needed for testing.
+
+
+def get_policy_command_description(policy_file):
+    return 'Check against policy {} using MediaConch'.format(policy_file)
