+++ features/steps/ccadipsteps.py
@@ -41,7 +41,7 @@ def step_impl(context):
     # Not the best way to check the AIP existence
     # as the request won't fail if AM is in debug mode.
     # Should we check directly in the SS API?
-    context.am_sel_cli.navigate_to_aip_in_archival_storage(context.aip_uuid)
+    context.am_user.amba.navigate_to_aip_in_archival_storage(context.aip_uuid)
 
 
 # Whens
